Issue 58: Bump ruby version to 2.4.2
Closes #58. 2.4.2 is the latest version of ruby [supported by Heroku](https://devcenter.heroku.com/articles/ruby-support#supported-runtimes). The current version is causing deploys of this app to Heroku to fail.

This also updates various dependencies in the lockfile using `bundle update #{name}` because of issues related to newer versions of ruby.

- `json` [⚠](https://github.com/flori/json/issues/286)
- `safe_yaml` [⚠](https://github.com/dtao/safe_yaml/issues/67)
- `webmock` [⚠] (https://github.com/bblimke/webmock/issues/683)

For folks that have this checked out locally, I also had to `brew install --force` to force an up to date build of `http_parse.rb`, which was throwing `Symbol not found: _rb_cFixnum`.

Finally, this updates petharbor cassette to include URL encoded quotes - this is probably just an internal change to `URI.encode_www_form`.